Add files via upload
The problems are  below: 
pb1: Find even and odd numbers from 1 to 100;
pb3: Arrange the number from big to small
pb4. Find the product of a number
pb5: If a square arm is 4 then fill up the square with "*"
pb6 : Find the summation of first and last digit of a number
pb7: Count the digit of a number or string
pb8.0: Arrange the two numbers from small to big
pb8.1: Arrange the two numbers from big to small
pb8.2: Arrange the number 30,12,89 from big to small
pb9: Test the number. is it a square number or not
pb10: find the factorial of a number 5!

#dartproblemsolve
#dartproblem
#dartsolve